[refactor/#23] meeting analysis 도메인 패키지명 정리
