+++ app/domains/meeting_analysis/schemas.py
