--- routers/decision.py
@@ -1,3 +1,0 @@
-from app.domains.decision.router import router
-
-__all__ = ["router"]
