+++ routers/meeting_analysis.py
@@ -1,0 +1,3 @@
+from app.domains.meeting_analysis.router import router
+
+__all__ = ["router"]
