+++ services/meeting_analysis_extraction.py
@@ -1,4 +1,4 @@
-from app.domains.decision.services.extraction import (
+from app.domains.meeting_analysis.services.extraction import (
     APPLICATION_POLICY_PROMPT,
     GEMINI_MODEL,
     extract_meeting_analysis,
